Add "forgot password?" link
We use xdg-open here to rely on the OS mime preference.
This is preferable than letting GTK handle the link.

Another reason to bypass GTK link handling is that we're running as root.
We don't want to launch the browser in root mode, but in user mode,
so we call xdg-open with "su -c" for that.

Note that this link only works in real mode (i.e. not when using the ./test script)
as it needs to know the username.