Remember the username/password/gateway
We didn't do this until now because it was considered a potential security issue
to save this information and let it be accessible without admin privs.

It turns out all that information (including the password in clear) is already
available to the local user without admin privs. Although the interface file
isn't readable, the information is available via the openvpn UI configuration
screens (and programmatically probably via NM libs).